+++ Meraki.Api/Data/FloorPlanUpdateRequest.cs
@@ -16,7 +16,7 @@ public class FloorPlanUpdateRequest
 	/// The floor number of the floor within the building.
 	/// </summary>
 	[DataMember(Name = "floorNumber")]
-	public int? FloorNumber { get; set; }
+	public double? FloorNumber { get; set; }
 
 	/// <summary>
 	/// Gets or Sets Center
